varias cosas + tema8

+++ tema-07/ejercicios/05-menu_Usuario/mvc-proyect/views/perfil/partials/menu.php
@@ -9,10 +9,10 @@
                     <a class="nav-link" href="<?= URL ?>perfil/edit">Editar</a>
                 </li>
                 <li class="nav-item">
-                    <a class="nav-link" href="<?= URL ?>perfil/pass">Cambiar password</a>
+                    <a class="nav-link" href="<?= URL ?>perfil/pass">Cambiar contraseña</a>
                 </li>
                 <li class="nav-item">
-                    <a class="nav-link" href="<?= URL ?>perfil/delete">Eliminar</a>
+                    <a class="nav-link" href="<?= URL ?>perfil/delete" onclick="return confirm('¿Seguro que quieres borrar tu perfil?')">Eliminar</a>
                 </li>
         </div>
     </div>
